requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
curl_cffi==0.6.2
distro==1.9.0
duckduckgo_search==5.2.1
gTTS==2.3.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
jaraco.context==4.3.0
lxml==5.1.0
more-itertools==10.2.0
openai==1.3.3
orjson==3.10.0
Pillow==10.1.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
python-dotenv==1.0.1
python-telegram-bot==20.3
pytube==15.0.0
readability-lxml==0.8.1
redis==5.0.3
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spotipy==2.23.0
tenacity==8.2.2
tiktoken==0.5.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
whois==0.9.27
wolframalpha==5.0.0
xmltodict==0.13.0
